move avatar utils out of public dir
