+++ app/core/matter/mini-game.ts
@@ -12,7 +12,7 @@ import { FloatingItem } from "../../models/colyseus-models/floating-item"
 import Player from "../../models/colyseus-models/player"
 import { PokemonAvatarModel } from "../../models/colyseus-models/pokemon-avatar"
 import { Portal, SynergySymbol } from "../../models/colyseus-models/portal"
-import { getOrientation } from "../../public/src/pages/utils/utils"
+import { getOrientation } from "../../utils/orientation"
 import GameRoom from "../../rooms/game-room"
 import GameState from "../../rooms/states/game-state"
 import { Transfer } from "../../types"
