+++ app/public/src/game/components/pokemon-avatar.ts
@@ -6,7 +6,7 @@ import { throttle } from "../../../../utils/function"
 import { SOUNDS, playSound } from "../../pages/utils/audio"
 import store from "../../stores"
 import { showEmote } from "../../stores/NetworkStore"
-import { getAvatarSrc, getAvatarString } from "../../utils"
+import { getAvatarSrc, getAvatarString } from "../../../../utils/avatar"
 import GameScene from "../scenes/game-scene"
 import EmoteMenu from "./emote-menu"
 import LifeBar from "./life-bar"
