+++ app/public/src/pages/component/after/team.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 import { ArraySchema } from "@colyseus/schema"
 import { IPokemonRecord } from "../../../../../models/colyseus-models/game-record"
-import { getAvatarSrc } from "../../../utils"
+import { getAvatarSrc } from "../../../../../utils/avatar"
 import "./team.css"
 
 export default function Team(props: { team: IPokemonRecord[] | ArraySchema<IPokemonRecord> }) {
