+++ app/public/src/pages/component/game/game-player-loading.tsx
@@ -3,7 +3,7 @@ import { CircularProgressbar, buildStyles } from "react-circular-progressbar"
 import "react-circular-progressbar/dist/styles.css"
 import { useTranslation } from "react-i18next"
 import { IPlayer } from "../../../../../types"
-import { getAvatarSrc } from "../../../utils"
+import { getAvatarSrc } from "../../../../../utils/avatar"
 import { getGameScene } from "../../game"
 import { cc } from "../../utils/jsx"
 import "./game-player-loading.css"
