+++ app/public/src/pages/component/game/game-stage-info.tsx
@@ -14,7 +14,7 @@ import { PkmIndex } from "../../../../../types/enum/Pokemon"
 import { SynergyAssociatedToWeather } from "../../../../../types/enum/Weather"
 import { min } from "../../../../../utils/number"
 import { selectCurrentPlayer, useAppSelector } from "../../../hooks"
-import { getAvatarSrc, getPortraitSrc } from "../../../utils"
+import { getAvatarSrc, getPortraitSrc } from "../../../../../utils/avatar"
 import { addIconsToDescription } from "../../utils/descriptions"
 import { cc } from "../../utils/jsx"
 import SynergyIcon from "../icons/synergy-icon"
