+++ app/public/src/pages/component/meta-report/item-statistic.tsx
@@ -2,7 +2,7 @@ import React from "react"
 import { useTranslation } from "react-i18next"
 import { IItemsStatistic } from "../../../../../models/mongo-models/items-statistic"
 import { PkmIndex } from "../../../../../types/enum/Pokemon"
-import { getPortraitSrc } from "../../../utils"
+import { getPortraitSrc } from "../../../../../utils/avatar"
 
 export default function ItemStatistic(props: {
   item: IItemsStatistic
