+++ app/public/src/pages/component/wiki/wiki-pokemons.tsx
@@ -12,7 +12,7 @@ import { Rarity } from "../../../../../types/enum/Game"
 import { Pkm, PkmFamily, PkmIndex } from "../../../../../types/enum/Pokemon"
 import { IPokemonData } from "../../../../../types/interfaces/PokemonData"
 import { groupBy } from "../../../../../utils/array"
-import { getPortraitSrc } from "../../../utils"
+import { getPortraitSrc } from "../../../../../utils/avatar"
 import { cc } from "../../utils/jsx"
 import { GamePokemonDetail } from "../game/game-pokemon-detail"
 import { PokemonTypeahead } from "../typeahead/pokemon-typeahead"
